Remove CentOS as docker image test
CentOS is no longer able to run a systemd based service from a docker
image.  It appears they were able to do so 5 years ago when the tests
were originally created, but the transition from init to systemd has
left them unwilling and unable to use systemd to control a service in
a docker image.

We'll need to find another way to test CentOS image installation.