Revert "Add OpenSUSE service check"
Services in a docker container and service checks on OpenSUSE are not
reasonable because they require systemd in much the same way that CentOS
requires systemd.

The test can install the rpm file and check that the rpm file is
installed correctly.  Checking that the service is running is a waste
of effort unless systemd is installed and used through the usual init
system for systemd.  That seems like too much adjusting, tweaking, and
guessing to be worth the trouble.  Let's find a different way to test
OpenSUSE services start and stop.

Whether it is the Vagrant system that Kohsuke used originally or some
other mechanism, it feels too heavyweight to try to guess all the
packages needed to have a systemd running and functional on an OpenSUSE
docker container.

This reverts commit 4a1a6821eb8c77877653c7ce656c180cd307a2ac.